☆Webアプリ/streamlit/requirements.txt

streamlit==1.29.0
Pillow==9.5.0
pandas==1.4.4
matplotlib==3.5.2


openpyxl==3.0.10

plotly==5.9.0
st-clickable-images==0.0.3

rembg==2.0.50
numba==0.57
numpy==1.23.0

folium
streamlit_folium

qrcode[pil]==7.3


seaborn

requests
bs4


moviepy
imageio

python-docx

gspread
oauth2client


streamlit.local_storage
pyocr
google-api-python-client

google-cloud-vision
pydrive2
icrawler
tensorflow

torch
torchvision

translate
transformers
sentencepiece

deepl-translate
mtcnn

gtts

mecab-python3

easyocr
unidic-lite
#ultralytics[core]
ultralytics
#opencv-python-headless
opencv-python==4.9.0.80

pytesseract

